+++ mods/gen1_wild_qol_nightly/CHANGELOG.md
@@ -7,6 +7,12 @@ was taken from.
 
 [stable]: https://github.com/wild1walker/Gen1WildQOL
 
+## [0.32.64] - 2026-09-05
+
+### Changed
+
+- Version moved with the nightly channel; nothing in this bundle changed.
+
 ## [0.32.63] - 2026-09-05
 
 ### Changed